# ddtrace/tracer: optimize ForEachBaggageItem.